feat: transfer slack emoji functions to incident module
